requirements.txt
appium-python-client==2.10.0
pytest==8.2.0
allure-pytest==2.13.2
